Ver. 1.0
This first version lays the groundwork for future builds. Specifically, from this point, there will be two forks. The first of which leans into the cat side of the program. The second fork leans into the data collection side of the program, opting to stay hidden. There is not yet an exe version of the .py file, however, it is intended and easily accomplishable.